fix: guard frame column NBT write and remove spurious cast in LiquidLevelDisplaySource
ReactorFrameDisplayManager.write() previously serialized frameColumnMinY and
frameColumnMaxY unconditionally, persisting their uninitialized default values
(0 / 0) for reactors that have no frame column set yet. Wrap the two putInt
calls in hasFrameColumn() so NBT is only written when the column bounds are
actually meaningful.

LiquidLevelDisplaySource cast the raw fluid amount to int via (int) before
assigning it to an int local variable; the explicit cast is redundant since
FluidStack.amount is already an int. Remove it to eliminate the unnecessary
narrowing-cast noise.

+++ src/main/java/net/nuclearteam/createnuclear/content/redstone/displayLink/source/LiquidLevelDisplaySource.java
@@ -25,7 +25,7 @@ public class LiquidLevelDisplaySource extends NumericSingleLineDisplaySource {
 
         int mode = context.sourceConfig().getInt("display_mode");
         var fluidList = controller.getBigFluidStack();
-        int fluid = (fluidList != null && !fluidList.isEmpty()) ? (int) fluidList.get(0).amount : 0;
+        int fluid = (fluidList != null && !fluidList.isEmpty()) ? fluidList.get(0).amount : 0;
         int maxFluid = 16000;
 
         return label.append(switch (mode) {
